Project Work Day 2
Added combat, monsterSpawning, playGame, and playerMovement methods. PlayerMovement is just taking inputs atm and not doing anything with them yet.

Need to find a way to put the player/monster objects into the map. (Maybe I need to create a new map that can hold objects not ints?)

But right now I can just have an Int as a placeholder where the player/monster is and compare where they are and move the player(int) around. If I just finish the project that way it will work but thats pretty ugly.